added check for whether pdf text exists

+++ gsp_pdf_reader.R
@@ -2,23 +2,29 @@
 #identifies the location of each section header
 #creates a folder of .txt files, where each file contains the text under each section
 
+library(stm)
 library(pdftools)
 
 pdfs <- list.files(path = "data_raw", pattern = "pdf", full.names = T)
-#pdf texts 1-51 successfully saved
-#category texts 1-51 successfully saved
+#pdf texts 1-112 successfully saved
+#category texts 1-112 successfully saved
 for(k in 1:length(pdfs)){
-   text <- pdf_text(pdfs[[k]])
-   print(k)
-   for(i in 1:length(text)){
-      if (nchar(text[i])> 10000){
-         text[i] <- NA
+   #if pdf doesn't exist, read in and save pdf text
+   if(!file.exists(paste0("data_output/",substr(pdfs[k],10,24),"_text"))){
+      text <- pdf_text(pdfs[[k]])
+      print(k)
+      for(i in 1:length(text)){
+         if (nchar(text[i])> 10000){
+            text[i] <- NA
+         }
       }
+      saveRDS(text, file = paste0("data_output/",substr(pdfs[k],10,24),"_text"))
+      #load gsp_xls_cleaner functions
+      page_key <- create_page_key(paste0(substr(pdfs[[k]],1,24),".xlsx"))
+      saveRDS(page_key, file = paste0("data_output/",substr(pdfs[k],10,24),"_categories"))
+      
    }
-   saveRDS(text, file = paste0("data_output/",substr(pdfs[k],10,24),"_text"))
-   page_key <- create_page_key(paste0(substr(pdfs[[k]],1,24),".xlsx"))
-   saveRDS(page_key, file = paste0("data_output/",substr(pdfs[k],10,24),"_categories"))
-   
+      
 }
 
 
@@ -28,11 +34,11 @@ gsp_num_id <- "0008"
 
 text_of_interest <- readRDS(paste0("data_output/gsp_num_id_",gsp_num_id,"_text"))
 key_of_interest <- readRDS(paste0("data_output/gsp_num_id_",gsp_num_id,"_categories"))
-
+View(key_of_interest)
 #TODO decide input text you want to have matched then go to page_vector of that element
 #words[key$page_vector[matched]]
 
-
+#TODO page contains substring of interest?
 #testocr <- pdf_ocr_text(file.path("data_raw/pdftools.pdf"))
 
 #TODO pdf_fonts to find sections
